Add variable read_only_permissions to state-manager-role
This is to support roles that can read from the state but cannot write to
it (duh) - which is for terraform_remote_state Data Source.

To make it possible the role permissions are split in rw and ro.
The rw policy is attached to the role only if read_only_permissions is
false.